graph_extractor/requirements.txt
pdf2image==1.17.0
pypdf==4.0.1
layoutparser==0.3.4
opencv-python==4.8.1.78
pytesseract==0.3.10
ollama==0.1.7
Pillow
PyYAML==6.0.1
numpy==1.24.3